--- custom_shell/ssd_writer.py
@@ -1,10 +1,0 @@
-import os
-
-
-class SSDWriter:
-    def write(self,
-              lba: int,
-              val: int) -> bool:
-        os.system(f"python ../ssd/ssd.py W {lba} {val}")
-
-        return True
